refactor(elasticache): extract free helper fns to helpers.rs submodule
Splits service.rs (5362 lines) by moving the free-fn helper block
(parse_*, collect_*, *_xml builders, paginate, parse_tags, etc.) from
the bottom of service.rs into a sibling `helpers.rs` module declared
via #[path].

Net effect:
- service.rs: 5362 -> 4227 (drops below the audit's 5000-line threshold)
- helpers.rs: 1142 (new file, all fns at `pub(crate)` visibility)

After this PR, only fakecloud-server/main.rs (5012 lines, single async
`main` fn) remains above 5000 in the workspace.

No behavior changes — fns moved verbatim, visibility widened from
private to pub(crate) so the parent module can see them via
`use helpers::*;`. All elasticache lib tests pass (177).
